Feat/add new modal diary

+++ Models/ViewModels/UpdateBudgetViewRequest.cs
@@ -3,7 +3,7 @@
     public class UpdateBudgetViewRequest
     {
         public int RubroTypeId { get; set; }
-        public int valorInicial { get; set; }
+        public decimal valorInicial { get; set; }
         public int Mes { get; set; }
         public int Anio { get; set; }
     }
